adds sections to their relevant plugin directories
